refactor: 🔨 Alterando o nome para "CardUser", e alterando o diretório do "CardUser".
